requirements.txt
fastapi
fastapi-cli
fastapi-mail
gunicorn
uvicorn 
pydantic 
sqlalchemy
passlib
bcrypt 
python-jose
python-multipart
postgres
alembic
redis
itsdangerous